% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_single_chapter_book.R
\name{buildSingleRepoBook}
\alias{buildSingleRepoBook}
\title{Build a gitbook from a skeleton repo and single package}
\usage{
buildSingleRepoBook(book_repo, book_path, pkg_repo)
}
\arguments{
\item{book_repo}{characterr(1) the URL of the book repo; this will
be cloned with git, so make sure that is exptected to work.}

\item{book_path}{character(1) Clone into this path}

\item{pkg_rep}{character(1) This is NOT a URL, but a string that 
is compatible with BiocManager::install.}
}
\description{
A desirable functionality when building a collaborative
bookdown book is the ability to build single chapters
without having to build all chapters. This function
needs a bookdown "book_repo" that contains the bookdown
skeleton and the repo (just the repo, not the URL) for
the "chapter package".
}
\details{
Steps involved are:
\itemize{
\item{Clone the book repo (must not exist)}
\item{Change into the book repo directory}
\item{Install (and capture output) the "chapter package", including
building the vignette locally.}
\item{Copy the installed Rmd file into the current directory}
\item{Build the pdf, epub, and gitbook}
\item{change back into the original directory}
}
}
