% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVignette.R
\name{checkVignette}
\alias{checkVignette}
\title{Check an `.Rmd` file that will be included in bookdown}
\usage{
checkVignette(rmdfile)
}
\arguments{
\item{rmdfile}{character(1) path to a `.Rmd` file}
}
\description{
Bookdown markdown headers are special in that each
chapter *must* start with a first-level header `#`.
}
\details{
A common workflow is to have each "chapter" respresented
as a single `.Rmd` file. Functions to check and then fix
`.Rmd` files programmatically facilitate finding and fixing
vignettes that do not meet these criteria.
}
