% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_utils.R
\name{cloneRepo}
\alias{cloneRepo}
\title{clone bookdown master repo}
\usage{
cloneRepo(...)
}
\arguments{
\item{...}{passed directly to git2r::clone}
}
\value{
a \code{\link[git2r]{repository}} object
}
\description{
clone bookdown master repo
}
\examples{
repo = cloneRepo('https://github.com/Bioconductor/BiocWorkshops2019', 'bioc2019')
repo$path
dirname(repo$path)

# find "HEAD"
git2r::repository_head(repo)

# latest commit
com = git2r::commits(repo)[[1]]
print(com$sha)
print(com$summary)

}
\seealso{
\code{\link[git2r]{clone}}
}
