% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_bib_files_utils.R
\name{collectBibliographyFiles}
\alias{collectBibliographyFiles}
\title{collect bibliograhy files from Rmd yaml frontmatter}
\usage{
collectBibliographyFiles(rmdfiles)
}
\arguments{
\item{rmdfiles}{character() vector of Rmd files
to scan for bibliography files}
}
\value{
character() vector of bibliography 
files from each of the Rmd files
}
\description{
collect bibliograhy files from Rmd yaml frontmatter
}
