% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_vignettes.R
\name{copy_vignettes}
\alias{copy_vignettes}
\title{copy vignettes from packages to staging directory}
\usage{
copy_vignettes(vignette_names, location = ".")
}
\arguments{
\item{vignette_names}{data.frame with the first column
containing the package and the second containing the name
of the associated RMD vignette (which should be in the `doc`
directory of the installed package)}

\item{location}{character(1) the name of the directory
into which to copy the vignettes}
}
\description{
This function takes as input a two-column data
frame containing the "package name" as the first
column and the "package vignette" in the second column.
The package vignette is expected to be in the `doc`
directory of the installed package.
}
\details{
In short, for each row in the `vignette_names` data 
frame, system.file locates the vignette named from 
the accompanying package and copies it to the "location"
directory
}
\examples{
vignette_names = read.csv(system.file('package_list.csv',package='BiocWorkshopPackageTest'), header=TRUE)
copy_vignettes(vignette_names)
dir()

}
