% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_rmd_without_yaml.R
\name{copyRmdWithoutYaml}
\alias{copyRmdWithoutYaml}
\title{copy an R markdown file, but remove yaml header}
\usage{
copyRmdWithoutYaml(srcfile, destfile = basename(srcfile))
}
\arguments{
\item{srcfile}{character(1) filename}

\item{destfile}{character(1) filename}
}
\description{
For building bookdown books, the markdown
header information usually present in an Rmd
file or a vignette causes problems. This 
function makes a copy of the Rmd, but removes
any yaml header information (specified by '^---' regex).
}
\details{
If srcfile and destfile are the same, the src file
is overwritten.
}
