% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInstalls.R
\name{figureInstallString}
\alias{figureInstallString}
\title{From git-like URLs, return the correct install string}
\usage{
figureInstallString(repos)
}
\arguments{
\item{repos}{character() vector of git URLs}
}
\description{
From git-like URLs, return the correct install string
}
\examples{

repos = c("https://github.com/seandavi/GEOquery",
          "https://git.bioconductor.org/packages/RNAseq123")
figureInstallString(repos)

}
