% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVignette.R
\name{installedVignettes}
\alias{installedVignettes}
\title{Check for installed Rmd files in system package directory}
\usage{
installedVignettes(pkgs)
}
\arguments{
\item{pkgs}{a character() vector of installed packages; 
\code{\link{installed.packages}} will be consulted
and uninstalled packages ignored}
}
\description{
When built packages are installed or packages are 
installed with building vignettes enabled, the vignette
`.Rmd` file(s) are included in the installation. This
function is just a convenience function for finding
package vignettes that have been installed, if any.
}
