% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookdown_file_utils.R
\name{linkSystemFiles}
\alias{linkSystemFiles}
\title{Link from absolute system.file to local directory}
\usage{
linkSystemFiles(..., destdir, include_full_path = TRUE)
}
\arguments{
\item{...}{passed to system.file}

\item{destdir}{character(1) the base destitation directory}

\item{include_full_path}{logical(1) should we include the actual path 
to the system.file() location.}
}
\description{
In the case that vignettes use system.file to
locate files for inclusion in the vignette, this
function can create a symbolic link that links
the `destdir` to the system.file(...) location.
}
\details{
For `include_full_path`, if system.file(...) is
 `/usr/abc/123`, if TRUE, the link will go from `/usr/abc/123` TO
 `destdir/usr/abc/123`. If FALSE, the link will go from `/usr/abc/123`
 to `destdir`.
}
