% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInstalls.R
\name{loggingPackageInstall}
\alias{loggingPackageInstall}
\title{install package and capture output}
\usage{
loggingPackageInstall(repo, capture_prefix = paste0(pkg, "-"),
  path = getwd())
}
\arguments{
\item{repo}{character(1) package name or repo (passed to BiocManager::install)}

\item{capture_prefix}{character(1) prefix of files that will end 
up as "capture_prefixstdout.log" and "capture_prefixstdout.log"}
}
\description{
Install a single package and capture stdout and stderr.
}
