% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInstalls.R
\name{reposFromURLs}
\alias{reposFromURLs}
\title{From git-like URLs, return the correct install string}
\usage{
reposFromURLs(urls)
}
\arguments{
\item{urls}{character() vector of git URLs}
}
\value{
named character() vector of repos that can be fed 
directly to \code{\link[BiocManager]{install}}
}
\description{
For github, uses \code{\link[remotes]{parse_github_url}} to do the
work. For Bioconductor, simply looks for stuff like:
"https://git.bioconductor.org/packages/PKGNAME".
}
\examples{

repos = c("https://github.com/seandavi/GEOquery",
          "https://git.bioconductor.org/packages/RNAseq123")
reposFromURLs(repos)

}
