% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInstalls.R
\name{testPackageInstalls}
\alias{testPackageInstalls}
\title{Test package installation success}
\usage{
testPackageInstalls(repos, update = FALSE, upgrade = "never",
  dependencies = TRUE, ...)
}
\arguments{
\item{repos}{character() character() vector of git URLs (that gets 
passed to `figureInstallString`)}

\item{update}{passed to \code{\link[BiocManager]{install}}}

\item{upgrade}{passed to \code{\link[BiocManager]{install}}}

\item{dependencies}{passed to \code{\link[BiocManager]{install}}}

\item{...}{passed to \code{\link[BiocManager]{install}}}
}
\description{
Test package installation success
}
\seealso{
\code{\link[remotes]{install_github}}
}
