% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_utils.R
\name{cloneRepo}
\alias{cloneRepo}
\title{clone bookdown master repo}
\usage{
cloneRepo(...)
}
\arguments{
\item{...}{passed directly to git2r::clone}
}
\value{
a \code{\link[git2r]{git_repository}} object
}
\description{
clone bookdown master repo
}
\seealso{
\code{\link[git2r]{clone}}
}
