% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_rmd_without_yaml.R
\name{copyRmdWithoutYaml}
\alias{copyRmdWithoutYaml}
\title{copy an R markdown file, optionally removing yaml frontmatter}
\usage{
copyRmdWithoutYaml(srcfile, destfile = basename(srcfile),
  remove.yaml = TRUE)
}
\arguments{
\item{srcfile}{character(1) filename}

\item{destfile}{character(1) filename}

\item{remove.yaml}{logical(1) when copying, remove yaml header
metadata?}
}
\description{
For building bookdown books, the markdown
header information usually present in an Rmd
file or a vignette causes problems. This 
function makes a copy of the Rmd, but removes
any yaml header information (specified by '^---' regex).
}
\details{
If srcfile and destfile are the same, the src file
is overwritten.
}
